back/requirements.txt
fastapi
fastapi-login
uvicorn[standard]
sqlalchemy
alembic
pymysql
jdatetime
pyjwt[crypto]
passlib[bcrypt]
python-multipart